split into files based on traits
