+++ src/double_ended_iterator.rs
@@ -1,0 +1,17 @@
+use pyo3::prelude::*;
+
+type PyDoubleEndedIteratorT =
+    Box<dyn DoubleEndedIterator<Item = PyResult<Py<PyAny>>> + Send + Sync>;
+#[pyclass]
+pub struct PyDoubleEndedIterator {
+    iter: PyDoubleEndedIteratorT,
+}
+
+impl PyDoubleEndedIterator {
+    pub fn rev<I>(iter: I) -> std::iter::Rev<I>
+    where
+        I: DoubleEndedIterator<Item = PyResult<Py<PyAny>>>,
+    {
+        iter.rev()
+    }
+}
