+++ src/exact_size_iterator.rs
@@ -1,0 +1,7 @@
+use pyo3::prelude::*;
+
+type PyExactSizeIteratorT = Box<dyn ExactSizeIterator<Item = PyResult<Py<PyAny>>> + Send + Sync>;
+#[pyclass]
+pub struct PyExactSizeIterator {
+    iter: PyExactSizeIteratorT,
+}
